realworld_deploy/robot_inference/requirements_robot.txt
# Robot 环境依赖 (Python 3.8, 兼容 Polymetis)
#
# 此环境用于运行 inference_client.py，通过轻量级 PolicyClient
# 与远程 GR00T 服务器通信，同时控制 Franka 机器人。
#
# 安装方法:
#   conda activate robot
#   pip install -r requirements_robot.txt
#
# 注意: Polymetis 需要单独安装 (参考 polymetis/ 目录)

# ZMQ 通信 (与 GR00T 服务器通信)
pyzmq>=22.0.0

# 消息序列化
msgpack>=1.0.0
msgpack-numpy>=0.4.7

# 数值计算
numpy>=1.19.0,<2.0.0

# PyTorch (Polymetis 依赖，需要与 Polymetis 版本匹配)
# torch>=1.9.0  # 已由 Polymetis 安装

# 相机支持
opencv-python>=4.5.0
pyrealsense2>=2.50.0  # Intel RealSense 相机 (可选)

# 可视化 (可选)
matplotlib>=3.3.0

